pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "onnxruntime>=1.21.1",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "rembg>=2.0.65",
    "reportlab>=4.4.0",
    "werkzeug>=3.1.3",
    "routes>=2.5.1",
    "openai>=1.75.0",
    "openpyxl>=3.1.5",
]
